Update the function of zscore
I have added the formula for zscore in a comment from line 3135 to 3137. If we consider the previous formula there is a huge difference in the output values of  that particular column. As  values are different from what we get from mean and standard deviation from Manuel calculation. Due to that mapping  to log and then going for mean and standard deviation there is a difference in values. Also, if doing mapping if a column have a nan values zscore is returning nan for a whole column, so if we fill the nan with mean and directly put the formula it will give better result.